They again changed their web site.
Unfortunately, the program schedule for different days is now on
different pages, which seem to have some hash value in their
filename part (of the URL.)  That means that for now, you can
only properly download today's episode (between 19:00 and 23:59:59
CE(S)T) with a proper filename.

  In theory, it seems I could get all names for the current month
or something like that by traversing the links within the HTML
file, but ...  Geee!